Consolodate document layout and page routes
